Fixed for 64bit compilation
git-svn-id: https://svn.aros.org/svn/aros/trunk@26457 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/prefs/Editor/JanoPrefs.c
@@ -84,7 +84,7 @@ struct TagItem *GadTags[]     = {IntTags, SepTags, TextFontTags, ScrFontTags, Sc
 /* Tag list for opening the main window */
 WORD  PrefZoom[4];
 struct TagItem PrefTags[] = {
-	{WA_Zoom,(ULONG)PrefZoom},
+	{WA_Zoom,PrefZoom},
 	{WA_NewLookMenus,TRUE},
 	{TAG_DONE}
 };
